split off stop fsm
