+++ src/software/ai/hl/stp/tactic/stop_fsm_test.cpp
@@ -1,0 +1,43 @@
+#include "software/ai/hl/stp/tactic/stop_fsm.h"
+
+#include <gtest/gtest.h>
+
+#include "software/test_util/test_util.h"
+
+TEST(StopFSMTest, test_transitions)
+{
+    World world = ::TestUtil::createBlankTestingWorld();
+    Robot robot(0,
+                RobotState(Point(1, -3), Vector(2.1, 3.1), Angle::half(),
+                           AngularVelocity::zero()),
+                Timestamp::fromSeconds(123));
+
+    boost::sml::sm<StopFSM> fsm;
+    EXPECT_TRUE(fsm.is(boost::sml::state<StopFSM::idle_state>));
+    fsm.process_event(StopFSM::Update{
+        .coast  = false,
+        .common = TacticUpdate{.robot      = robot,
+                               .world      = world,
+                               .set_intent = [](std::unique_ptr<Intent>) {}}});
+    // robot is still moving
+    EXPECT_TRUE(fsm.is(boost::sml::state<StopFSM::stop_state>));
+    robot = Robot(0,
+                  RobotState(Point(1, -3), Vector(1.1, 2.1), Angle::half(),
+                             AngularVelocity::zero()),
+                  Timestamp::fromSeconds(123));
+    fsm.process_event(StopFSM::Update{
+        .coast  = false,
+        .common = TacticUpdate{.robot      = robot,
+                               .world      = world,
+                               .set_intent = [](std::unique_ptr<Intent>) {}}});
+    // robot is still moving
+    EXPECT_TRUE(fsm.is(boost::sml::state<StopFSM::stop_state>));
+    robot = TestUtil::createRobotAtPos(Point(1, -3));
+    fsm.process_event(StopFSM::Update{
+        .coast  = false,
+        .common = TacticUpdate{.robot      = robot,
+                               .world      = world,
+                               .set_intent = [](std::unique_ptr<Intent>) {}}});
+    // robot stopped
+    EXPECT_TRUE(fsm.is(boost::sml::X));
+}
