+++ src/software/ai/hl/stp/tactic/stop_tactic.h
@@ -1,45 +1,9 @@
 #pragma once
 
 #include "software/ai/hl/stp/action/move_action.h"  // TODO (#1888): remove this dependency
+#include "software/ai/hl/stp/tactic/stop_fsm.h"
 #include "software/ai/hl/stp/tactic/tactic.h"
 #include "software/ai/intent/stop_intent.h"
-
-struct StopFSM
-{
-    class idle_state;
-    class stop_state;
-
-    struct Update
-    {
-        bool coast;
-        TacticUpdate common;
-    };
-
-    auto operator()()
-    {
-        using namespace boost::sml;
-
-        const auto idle_s   = state<idle_state>;
-        const auto stop_s   = state<stop_state>;
-        const auto update_e = event<Update>;
-
-        const auto update_stop = [](auto event) {
-            event.common.set_intent(
-                std::make_unique<StopIntent>(event.common.robot.id(), event.coast));
-        };
-
-        const auto stop_done = [](auto event) {
-            return robotStopped(event.common.robot);
-        };
-
-        return make_transition_table(
-            // src_state + event [guard] / action = dest state
-            *idle_s + update_e / update_stop            = stop_s,
-            stop_s + update_e[!stop_done] / update_stop = stop_s,
-            stop_s + update_e[stop_done] / update_stop  = X,
-            X + update_e[!stop_done] / update_stop      = X);
-    }
-};
 
 /**
  * The StopTactic will stop the robot from moving. The robot will actively try and brake
