+++ src/software/ai/hl/stp/tactic/stop_tactic_test.cpp
@@ -64,41 +64,3 @@ TEST(StopTacticTest, test_calculate_robot_cost)
     // equally
     EXPECT_EQ(0.5, tactic.calculateRobotCost(robot, world));
 }
-
-TEST(MoveFSMTest, test_transitions)
-{
-    World world = ::TestUtil::createBlankTestingWorld();
-    Robot robot(0,
-                RobotState(Point(1, -3), Vector(2.1, 3.1), Angle::half(),
-                           AngularVelocity::zero()),
-                Timestamp::fromSeconds(123));
-
-    boost::sml::sm<StopFSM> fsm;
-    EXPECT_TRUE(fsm.is(boost::sml::state<StopFSM::idle_state>));
-    fsm.process_event(StopFSM::Update{
-        .coast  = false,
-        .common = TacticUpdate{.robot      = robot,
-                               .world      = world,
-                               .set_intent = [](std::unique_ptr<Intent>) {}}});
-    // robot is still moving
-    EXPECT_TRUE(fsm.is(boost::sml::state<StopFSM::stop_state>));
-    robot = Robot(0,
-                  RobotState(Point(1, -3), Vector(1.1, 2.1), Angle::half(),
-                             AngularVelocity::zero()),
-                  Timestamp::fromSeconds(123));
-    fsm.process_event(StopFSM::Update{
-        .coast  = false,
-        .common = TacticUpdate{.robot      = robot,
-                               .world      = world,
-                               .set_intent = [](std::unique_ptr<Intent>) {}}});
-    // robot is still moving
-    EXPECT_TRUE(fsm.is(boost::sml::state<StopFSM::stop_state>));
-    robot = TestUtil::createRobotAtPos(Point(1, -3));
-    fsm.process_event(StopFSM::Update{
-        .coast  = false,
-        .common = TacticUpdate{.robot      = robot,
-                               .world      = world,
-                               .set_intent = [](std::unique_ptr<Intent>) {}}});
-    // robot stopped
-    EXPECT_TRUE(fsm.is(boost::sml::X));
-}
